requirements.txt
backcall==0.2.0
cycler==0.10.0
decorator==4.4.2
imageio==2.9.0
ipdb==0.13.4
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
kiwisolver==1.3.1
matplotlib==3.3.3
networkx==2.5
numpy==1.19.4
pandas==1.1.5
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==3.0.8
ptyprocess==0.6.0
Pygments==2.7.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.4
six==1.15.0
tifffile==2020.9.3
traitlets==4.3.3
wcwidth==0.2.5
